Rename stellarBlockchain folders to stellar
