+++ common/enum/stellar/StellarPlatformEnum.ts
