+++ common/errors/stellar/NoAssetSelectedError.ts
