+++ common/interfaces/stellar/IAdditionalPaymentFilter.ts
