+++ common/interfaces/stellar/IAsset.ts
