+++ common/interfaces/stellar/IAssetsPath.ts
