+++ common/interfaces/stellar/ILiquidityPool.ts
