+++ common/interfaces/stellar/ILiquidityPoolPrice.ts
