+++ common/interfaces/stellar/ILiquidityPoolReserve.ts
