+++ common/interfaces/stellar/ILiquidtyPoolPrice.ts
