+++ common/interfaces/stellar/IPredicate.ts
