+++ common/interfaces/stellar/IPriceFraction.ts
