+++ common/interfaces/stellar/ISigner.ts
