+++ common/types/stellar/RevokeSponsorshipType.ts
