+++ common/utils/stellar/convertAmountToBigNumber.ts
