+++ nodes/Stellar/actions/liquidityPool/liquidityPoolDeposit/execute.ts
@@ -1,8 +1,8 @@
 import { IExecuteFunctions } from 'n8n-workflow';
 import { Operation } from 'stellar-sdk';
-import { convertAmountToBigNumber } from '../../../../../common/utils/stellarBlockchain/convertAmountToBigNumber';
+import { convertAmountToBigNumber } from '../../../../../common/utils/stellar/convertAmountToBigNumber';
 
-import ILiquidityPoolPrice from '../../../../../common/interfaces/stellarBlockchain/ILiquidityPoolPrice';
+import ILiquidityPoolPrice from '../../../../../common/interfaces/stellar/ILiquidityPoolPrice';
 import getPrice from './helpers/helpers';
 
 export async function liquidityPoolDeposit(this: IExecuteFunctions) {
