+++ nodes/Stellar/actions/liquidityPool/liquidityPoolDeposit/helpers/__tests__/helpers.test.ts
@@ -3,7 +3,7 @@
  */
 import { describe, expect, it } from 'vitest';
 import getPrice from '../helpers';
-import ILiquidityPoolPrice from '../../../../../../../common/interfaces/stellarBlockchain/ILiquidityPoolPrice';
+import ILiquidityPoolPrice from '../../../../../../../common/interfaces/stellar/ILiquidityPoolPrice';
 import NoPriceSelectedError from '../../errors/NoPriceSelectedError';
 
 describe('Get liquidity pool price', () => {
