+++ nodes/Stellar/actions/offers/manageSellOffer/execute.ts
@@ -1,9 +1,9 @@
 import { IExecuteFunctions } from 'n8n-workflow';
 import { Operation, Asset } from 'stellar-sdk';
-import { convertAmountToBigNumber } from '../../../../../common/utils/stellarBlockchain/convertAmountToBigNumber';
-import IAsset from '../../../../../common/interfaces/stellarBlockchain/IAsset';
-import { buildAsset } from '../../../../../common/utils/stellarBlockchain/buildAsset';
-import { StellarPlatformEnum } from '../../../../../common/enum/stellarBlockchain/StellarPlatformEnum';
+import { convertAmountToBigNumber } from '../../../../../common/utils/stellar/convertAmountToBigNumber';
+import IAsset from '../../../../../common/interfaces/stellar/IAsset';
+import { buildAsset } from '../../../../../common/utils/stellar/buildAsset';
+import { StellarPlatformEnum } from '../../../../../common/enum/stellar/StellarPlatformEnum';
 
 export async function manageSellOffer(this: IExecuteFunctions) {
 	try {
