+++ nodes/Stellar/actions/server/findPayments/helpers/helpers.ts
@@ -1,7 +1,7 @@
 import { Server, ServerApi } from 'stellar-sdk';
 import Payment from '../../../entities/Payment';
 import NoPaymentFoundError from '../errors/NoPaymentFoundError';
-import IAsset from '../../../../../../common/interfaces/stellarBlockchain/IAsset';
+import IAsset from '../../../../../../common/interfaces/stellar/IAsset';
 
 function paymentMapper(paymentRecord: ServerApi.PaymentOperationRecord) {
 	const {
